Use mypy
This commit updates our CI to do type checking with mypy. It includes a
few changes in order to get the checks to pass. Also, we are now running
`uv ruff check` for linting.

The `.python-version` file was removed since we are now relying on `uv`
to manage the python version.