Frame JSON-LD as part of event
This commit updates the JSON-LD framing to take place in a SQLAlchemy
set event [1] prior to when the JSON-LD is persisted to the database for
Work and Instances (but not OtherResources).

This means other parts of the codebase no longer need to remember to
frame the JSON-LD. This is nice because other parts of the code working
with Work and Instance objects can now rely on the data having been
framed, and can work with it as a JSON data structure if necessary.

The type signature of `utils.graph.frame_jsonld()` has been changed to
take a Python data structure as a parameter, rather than an rdflib
Graph. This was done so that we no longer need to serialize the Graph as
a JSON-LD str, and then parse it, in order to frame it.

NOTE: the use of frame_jsonld in bluecore-workflows and bluecore_api
will need to be removed when they are upgraded to use bluecore-models v0.8.0.

[1] https://docs.sqlalchemy.org/en/20/orm/events.html#sqlalchemy.orm.AttributeEvents.set