Better JSON-LD framing errors
When `Work.data` or `Instance.data` are set a SQLAlchemy event ensures that
the JSON-LD data is "framed". For it to do this the .uri property must
also be set.

We had some cases where data was being set before uri,
which was causing a very difficult to interpret error message. This
commit adds an exception to help better understand the error.

This works fine (the order of the properties matters):

```
Work(uri="https://example.com/", data={...})
```

but this does not:

```
Work(data={...}, uri="https://example.com/")

ValueError: For automatic jsonld framing to work you must ensure the uri property is set before the data property, even when constructing an object.
```