Prevent duplicate versions
This is a branch of the `t57-version-fix` branch in #60 which provides alternate logic for determining whether to create a new version using a technique suggested in the [SQLAlchemy docs](https://github.com/blue-core-lod/bluecore-models/pull/60/commits/83648d8dc83194f9bce5ace82421782f26cb25af):

> This method is called for all instances that are marked as “dirty”, even those which have no net changes to their column-based attributes, and for which no UPDATE statement has proceeded. An object is marked as dirty when any of its column-based attributes have a “set attribute” operation called or when any of its collections are modified. If, at update time, no column-based attributes have any net changes, no UPDATE statement will be issued. This means that an instance being sent to MapperEvents.after_update() is not a guarantee that an UPDATE statement has been issued.
>
> To detect if the column-based attributes on the object have net changes, and therefore resulted in an UPDATE statement, use object_session(instance).is_modified(instance, include_collections=False).

It also introduces a commit after the new Work is created in the new test. Without this, the new version will never get created even if there is no logic in place to prevent it.

Fixes #57