Experimenting with GraphMap
This commit is here to roll back to in case I want to bring back the
approach of using a GraphMap dictionary. I think instead I'm going to
try introducing a BluecoreGraph class for managing the partitioning and
saving of objects since there's so much state that needs to be shared
between the different functions.