Add BluecoreGraph class (#70)
* Add save_graph() and partition_graph()

The new `save_graph()` function is intended to make it easier to test
that graph data is being inserted and updated in the database correctly.

In order to do this logic that was previously available in bluecore-workflows for extracting works, instances and other resources from a larger graph has been ported to the new `partition_graph()` function.

Refs https://github.com/blue-core-lod/bluecore-workflows/issues/67

* New version