Do not frame "profile" Other Resources
This commmit updates the automatic json-ld framing so that it does not
run for OtherResource where is_profile is set to True. This is needed
because "profile" OtherResources don't store json-ld in their data
property, and framing it that way will cause problems.

Since we are framing the json-ld for Work, Instance and OtherResource
objects the functionality and event listener were moved to BaseResource.