Address linking issues
This commit adjusts the derivedFrom lookup to look in the graph being
saved as well as the database for a pre-existing Blue Core URI to use.

It also adds rdf:type assertions using bf:hasInstance and bf:instanceOf
assertions, which may omit the type, and confuse the linking logic
(which expects to find them).

fixes #75

There is some follow up work to ensure that the bluecore-api uses
BluecoreGraph to persist to the database and ensure that the URI minting
behavior is consistent: https://github.com/blue-core-lod/bluecore_api/issues/193